Expose HMC model in KB0371 notebook
